+++ src/Controller/VeterinarianController.php
@@ -6,6 +6,7 @@ use App\Entity\Veterinarian;
 use App\Form\VeterinarianType;
 use App\Repository\VeterinarianRepository;
 use Knp\Component\Pager\PaginatorInterface;
+use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -38,16 +39,26 @@ class VeterinarianController extends AbstractController
     /**
      * @Route("/new", name="app_veterinarian_new", methods={"GET", "POST"})
      */
-    public function new(Request $request, VeterinarianRepository $veterinarianRepository): Response
+    public function new(Request $request, VeterinarianRepository $veterinarianRepository, LoggerInterface $logger): Response
     {
         $veterinarian = new Veterinarian();
         $form = $this->createForm(VeterinarianType::class, $veterinarian);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $veterinarianRepository->add($veterinarian, true);
 
-            return $this->redirectToRoute('app_veterinarian_index', [], Response::HTTP_SEE_OTHER);
+            $farms = $form->get('farms')->getData();
+            foreach ($farms as $farm) {
+                $veterinarian->addFarm($farm);
+            }
+
+            try {
+                $veterinarianRepository->add($veterinarian, true);
+                $this->addFlash('success', 'Veterinario adicionado com sucesso.');
+                return $this->redirectToRoute('app_veterinarian_index', [], Response::HTTP_SEE_OTHER);
+            } catch (\Exception $e) {
+                $this->addFlash('error', 'Erro ao criar veterinário: ' . $e->getMessage());
+            }
         }
 
         return $this->renderForm('veterinarian/new.html.twig', [
@@ -55,6 +66,7 @@ class VeterinarianController extends AbstractController
             'form' => $form,
         ]);
     }
+
 
     /**
      * @Route("/{id}", name="app_veterinarian_show", methods={"GET"})
@@ -75,9 +87,13 @@ class VeterinarianController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $veterinarianRepository->add($veterinarian, true);
-
-            return $this->redirectToRoute('app_veterinarian_index', [], Response::HTTP_SEE_OTHER);
+            try {
+                $veterinarianRepository->add($veterinarian, true);
+                $this->addFlash('success', 'Veterinario editado com sucesso.');
+                return $this->redirectToRoute('app_veterinarian_index', [], Response::HTTP_SEE_OTHER);
+            } catch (\Exception $e) {
+                $this->addFlash('error', 'Erro ao editar veterinário: ' . $e->getMessage());
+            }
         }
 
         return $this->renderForm('veterinarian/edit.html.twig', [
@@ -89,10 +105,24 @@ class VeterinarianController extends AbstractController
     /**
      * @Route("/{id}", name="app_veterinarian_delete", methods={"POST"})
      */
-    public function delete(Request $request, Veterinarian $veterinarian, VeterinarianRepository $veterinarianRepository): Response
+    public function delete($id, Request $request, VeterinarianRepository $veterinarianRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$veterinarian->getId(), $request->request->get('_token'))) {
-            $veterinarianRepository->remove($veterinarian, true);
+        $veterinarian = $veterinarianRepository->find($id);
+
+        if (!$veterinarian) {
+            $this->addFlash('error', 'Veterinario não encontrado.');
+            return $this->redirectToRoute('app_veterinarian_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        try {
+            if ($this->isCsrfTokenValid('delete' . $veterinarian->getId(), $request->request->get('_token'))) {
+                $veterinarianRepository->remove($veterinarian, true);
+                $this->addFlash('success', 'Veterinario deletado com sucesso.');
+            } else {
+                throw new \Exception('Erro ao deletar.');
+            }
+        } catch (\Exception $e) {
+            $this->addFlash('error', 'Erro ao deletar. ' . $e->getMessage());
         }
 
         return $this->redirectToRoute('app_veterinarian_index', [], Response::HTTP_SEE_OTHER);
